create popup function to for reusing create retrive item function to make it easy to reuse
