+++ src/display-item.js
@@ -1,8 +1,10 @@
-import retrieveItem from "./retrieve-item";
+import popUp from './pop-up.js';
+import retrieveItem from './retrieve-item.js';
 
-const displayItem = async (item) => {
-    const displayItem = await retrieveItem(item);
-    console.log(displayItem);
-}
+const displayItem = async (id) => {
+  const displayItem = await retrieveItem(id);
+  const charData = displayItem.data.results[0];
+  popUp(charData);
+};
 
 export default displayItem;
