+++ src/index.html
@@ -20,10 +20,6 @@
     </header>
     <main>
         <div class="d-flex super-heroes-container"></div>
-        <div class="display-item">
-            <img src="" alt="">
-            <h2 class="item-title">Dummy text</h2>
-        </div>
     </main>
     <footer id="footer">
         Marvel's Characters by Alex, Abel & Jim
