+++ src/retrieve-item.js
@@ -1,9 +1,6 @@
-import apiCall from "./marvel-api-call";
-
-const retrieveItem = async (item) => {
-    const data = await apiCall();
-    const result = data.filter(hero => hero.name === item.children[1].children[0].innerHTML);
-    return result;
-}
+const retrieveItem = async (id) => {
+  const character = await fetch(`https://gateway.marvel.com/v1/public/characters/${id}?ts=1&hash=cef440fde913b8756e1db19d5f18cd9e&apikey=580e7d5773a2965034fd7fcd3cad5f53`);
+  return character.json();
+};
 
 export default retrieveItem;
